+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
 CHANGELOG
 ---------
+
+Unreleased
+::::::::::
+
+- [`fixed`] Fix typo: Carbon Dioxide
+
+
 0.1.0
 :::::
 - Initial release